rename PromotedPropertyResolver to PromotedPropertyCandidateResolver
